sentiment_analysis/requirement.txt
tweepy
mysql-connector-python
nltk
scikit-learn
pandas
pickle-mixin
